Check Authorization header in logout. Delete all sessions if specified

+++ src/main/java/ule/chat/controllers/AuthController.java
@@ -2,8 +2,11 @@ package ule.chat.controllers;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.gson.JsonObject;
+import com.google.gson.stream.JsonReader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.IOUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -15,6 +18,7 @@ import ule.chat.service.SessionService;
 import ule.chat.service.UserService;
 import ule.chat.utils.Constants;
 
+import javax.servlet.ServletInputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -63,17 +67,33 @@ public class AuthController {
 	}
 
 	@PostMapping(Routes.LOGOUT_URL)
-	public void logout(HttpServletRequest request) {
-		// Todo: check if we must delete all sessions. If only one session is requested,
-		//  we take the token, and we delete the session. If more than one session is
-		//  requested, we take the username, and we delete all the sessions.
-		String token = request.getHeader(AUTHORIZATION).substring(JWT_TOKEN_PREFIX.length());
+	public void logout(HttpServletRequest request, HttpServletResponse response) throws IOException {
+		String authorizationHeader = request.getHeader(AUTHORIZATION);
+
+		if (authorizationHeader == null) { // If request does not contain authorization header send error.
+			response.setStatus(HttpStatus.BAD_REQUEST.value());
+			response.setContentType(APPLICATION_JSON_VALUE);
+			new ObjectMapper().writeValue(response.getWriter(), "You must provide the authorization token");
+			return;
+		}
+
+		String token = authorizationHeader.substring(JWT_TOKEN_PREFIX.length());
 		DecodedJWT decodedJWT = JWT_VERIFIER.verify(token);
 		String username = decodedJWT.getSubject();
 
-		new SecurityContextLogoutHandler().logout(request, null, null);
-
 		log.info("User {} has been logged out", username);
-		this.sessionService.deleteAllSessionsOfUser(username);
+
+		String body = IOUtils.toString(request.getReader());
+		JsonObject jsonBody = Constants.GSON.fromJson(body, JsonObject.class);
+
+		this.sessionService.deleteSession(token); // Always delete the session.
+
+		// And then check if we must delete all the other sessions.
+		if (jsonBody.has("deleteAllSessions")) {
+			this.sessionService.deleteAllSessionsOfUser(username);
+		}
+
+		// Logout from Spring.
+		new SecurityContextLogoutHandler().logout(request, null, null);
 	}
 }
